Try a workaround for tests which started failing with new pytest version (>= 8.0).
In theory with mocking in place, those tests should be safe to run in
parallel, but because of some reason, they are prone to race condition,
but only with pytest >= 8.0.0.

To avoid this race condition, we now run tests which are not thread safe
separately at the end in non parallel fashion.

Tests which are not safe to run in parallel are marked with "serial"
pytest marker.